update_tracking() moved to class Room
