+++ room.py
@@ -1,8 +1,10 @@
 from player import Player
-from monster import Vampire, Hunter, Zombie
+from monster import Vampire, Hunter, Zombie, Monster
 from item import SpeedBoost, Fog
 from stationary_things import Wall, Exit, Start, Floor
+from collections import namedtuple
 
+Point = namedtuple('Point', ['y', 'x'])
 
 class Room:
     def __init__(self, square_size, width=0, height=0):
@@ -92,3 +94,23 @@ class Room:
         self.height = len(self.static_map)
         self.width = col_count
         self.tracking_map = [[None] * self.width for y in range(self.height)]
+
+    def update_tracking(self):
+        """
+        Low-level function called every frame in order to update room.tracking_map via BFS algorithm.
+        """
+        self.tracking_map = [[None] * self.width for y in range(self.height)]
+        visited = set()
+        queue = [(Point(self.player.y, self.player.x), 0)]
+        while queue:
+            node, level = queue.pop(0)
+            if node not in visited:
+                visited.add(node)
+                y, x = node.y, node.x
+                self.tracking_map[y][x] = level
+                for neighbour in (Point(y - 1, x),
+                                  Point(y + 1, x),
+                                  Point(y, x - 1),
+                                  Point(y, x + 1)):
+                    if Monster.class_can_move_to(neighbour.x, neighbour.y, self) and neighbour not in visited:
+                        queue.append((neighbour, level + 1))
